info
This short piece of code has been previously used for a project in my first programming class, where I learned Python. The name of the game is CryptoLogic. The idea of the game is to scramble a random word from a file called wordlist.txt which is not included because this can be any list of words separated by a space in a normal text file. The first few lines of code are to used to open that file as r to be used in the next few lines of code. Those next lines import Python's shuffle function, creates the secret word that is pulled from wordlist.txt randomly, then randomly scrambles the word into mixed up letters. Following there are a word counter which counts the amount of correct and incorrect guesses you do. After this is followed by a series of print messages that welcomes the user to the game and explains how things work. The next loop statement is the main sequence, which counts the incorrect guesses and tells you the letters you have correct in order. Once you go through and finally guess all of the correct letters, you are met with a print statement that congratulates you on winning, and tells you what the correct word is altogether.